requirements.txt
bikram==1.0.0
Django==2.2.3
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
